Cleaned up part of camera.py and got ready to turn it into a library code
Created the visca_over_ip directory which will contain my library code
